+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Breaking Changes
 
+## v0.1.2
+
+### Promises introduced
+
+The following APIs that took in a callback function as a parameter now instead return a Promise.
+
+* Location.getLocation
+* Location.showLocation
+
 ## v0.1.1
 
 ### Moved APIs